Factor out BehaviorNodeContainer to container.rs
