+++ src/lib.rs
@@ -666,6 +666,7 @@
 //! * [ ] Static type checking for behavior tree definition file
 //!
 
+mod container;
 mod context;
 pub mod error;
 mod nodes;
@@ -678,6 +679,7 @@ use std::any::Any;
 use std::collections::HashMap;
 use std::rc::Rc;
 
+pub use crate::container::BehaviorNodeContainer;
 pub use crate::context::Context;
 pub use crate::nodes::{tick_child_node, FallbackNode, SequenceNode};
 pub use crate::symbol::Symbol;
@@ -687,7 +689,6 @@ pub use crate::{
     registry::{boxify, Constructor, Registry},
 };
 pub use ::once_cell::sync::*;
-use error::{AddChildError, AddChildResult};
 
 #[derive(PartialEq, Eq, Debug, Clone, Copy)]
 pub enum BehaviorResult {
@@ -772,59 +773,6 @@ pub trait BehaviorNode {
 
     fn num_children(&self) -> NumChildren {
         NumChildren::Finite(0)
-    }
-}
-
-pub struct BehaviorNodeContainer {
-    node: Box<dyn BehaviorNode>,
-    blackboard_map: HashMap<Symbol, BlackboardValue>,
-    child_nodes: Vec<BehaviorNodeContainer>,
-}
-
-impl BehaviorNodeContainer {
-    pub fn new(
-        node: Box<dyn BehaviorNode>,
-        blackboard_map: HashMap<Symbol, BlackboardValue>,
-    ) -> Self {
-        Self {
-            node,
-            blackboard_map,
-            child_nodes: vec![],
-        }
-    }
-
-    pub fn new_raw(node: Box<dyn BehaviorNode>) -> Self {
-        Self {
-            node,
-            blackboard_map: HashMap::new(),
-            child_nodes: vec![],
-        }
-    }
-
-    pub fn new_node(node: impl BehaviorNode + 'static) -> Self {
-        Self {
-            node: Box::new(node),
-            blackboard_map: HashMap::new(),
-            child_nodes: vec![],
-        }
-    }
-
-    pub fn tick(&mut self, arg: BehaviorCallback, ctx: &mut Context) -> BehaviorResult {
-        std::mem::swap(&mut self.child_nodes, &mut ctx.child_nodes.0);
-        std::mem::swap(&mut self.blackboard_map, &mut ctx.blackboard_map);
-        let res = self.node.tick(arg, ctx);
-        std::mem::swap(&mut self.blackboard_map, &mut ctx.blackboard_map);
-        std::mem::swap(&mut self.child_nodes, &mut ctx.child_nodes.0);
-        res
-    }
-
-    pub fn add_child(&mut self, child: BehaviorNodeContainer) -> AddChildResult {
-        if NumChildren::Finite(self.child_nodes.len()) < self.node.num_children() {
-            self.child_nodes.push(child);
-            Ok(())
-        } else {
-            Err(AddChildError::TooManyNodes)
-        }
     }
 }
